+++ tests/TestCase/Shell/InstallShellTest.php
@@ -1,0 +1,139 @@
+<?php
+
+/**
+ * This file is part of me-cms-youtube.
+ *
+ * me-cms-youtube is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * me-cms-youtube is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with me-cms-youtube.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * @author      Mirko Pagliai <mirko.pagliai@gmail.com>
+ * @copyright   Copyright (c) 2016, Mirko Pagliai for Nova Atlantis Ltd
+ * @license     http://www.gnu.org/licenses/agpl.txt AGPL License
+ * @link        http://git.novatlantis.it Nova Atlantis Ltd
+ * @see         MeCms\Utility\SitemapBuilder
+ */
+namespace MeCmsYoutube\Test\TestCase\Shell;
+
+use Cake\Console\ConsoleIo;
+use Cake\TestSuite\Stub\ConsoleOutput;
+use Cake\TestSuite\TestCase;
+use MeCmsYoutube\Shell\InstallShell;
+use Reflection\ReflectionTrait;
+
+/**
+ * InstallShellTest class
+ */
+class InstallShellTest extends TestCase
+{
+    use ReflectionTrait;
+
+    /**
+     * @var \MeCmsInstagram\Shell\InstallShell
+     */
+    protected $InstallShell;
+
+    /**
+     * @var \Cake\TestSuite\Stub\ConsoleOutput
+     */
+    protected $err;
+
+    /**
+     * @var \Cake\TestSuite\Stub\ConsoleOutput
+     */
+    protected $out;
+
+    /**
+     * Setup the test case, backup the static object values so they can be
+     * restored. Specifically backs up the contents of Configure and paths in
+     *  App if they have not already been backed up
+     * @return void
+     */
+    public function setUp()
+    {
+        parent::setUp();
+
+        $this->out = new ConsoleOutput();
+        $this->err = new ConsoleOutput();
+        $io = new ConsoleIo($this->out, $this->err);
+        $io->level(2);
+
+        $this->InstallShell = $this->getMockBuilder(InstallShell::class)
+            ->setMethods(['_stop', 'copyConfig'])
+            ->setConstructorArgs([$io])
+            ->getMock();
+    }
+
+    /**
+     * Teardown any static object changes and restore them
+     * @return void
+     */
+    public function tearDown()
+    {
+        parent::tearDown();
+
+        unset($this->InstallShell, $this->err, $this->out);
+    }
+
+    /**
+     * Test for `__construct()` method
+     * @test
+     */
+    public function testConstruct()
+    {
+        $this->assertEquals([
+            'MeCmsYoutube.me_cms_youtube',
+        ], $this->getProperty($this->InstallShell, 'config'));
+    }
+
+    /**
+     * Test for `all()` method
+     * @test
+     */
+    public function testAll()
+    {
+        $this->InstallShell->method('copyConfig')
+            ->will($this->returnCallback(function () {
+                $this->out->write('called `copyConfig`');
+            }));
+
+        //Calls with `force` options
+        $this->InstallShell->params['force'] = true;
+        $this->InstallShell->all();
+
+        //Calls with no interactive mode
+        unset($this->InstallShell->params['force']);
+        $this->InstallShell->interactive = false;
+        $this->InstallShell->all();
+
+        $this->assertEquals([
+            'called `copyConfig`',
+            'called `copyConfig`',
+        ], $this->out->messages());
+        $this->assertEmpty($this->err->messages());
+    }
+
+    /**
+     * Test for `getOptionParser()` method
+     * @test
+     */
+    public function testGetOptionParser()
+    {
+        $parser = $this->InstallShell->getOptionParser();
+
+        $this->assertEquals('Cake\Console\ConsoleOptionParser', get_class($parser));
+        $this->assertEquals([
+            'all',
+            'copyConfig',
+        ], array_keys($parser->subcommands()));
+    }
+}