+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * `Youtube::getInfo()` method returns an object;
 * fixed a bug with videos longer than an hour;
 * renamed repository and package. Now is `me-cms-youtube`;
-* added test for `Youtube` class.
+* added test for `Youtube` and `InstallShell` classes.
 
 ## 2.6 branch
 ### 2.6.7